renamed dogProfileForm to have consistent naming scheme
