+++ client/src/pages/DogCreeationForm/DogCreationForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './DogProfileForm.css'
+import './DogCreationForm.css'
 import axios from 'axios'
 import env from 'react-dotenv'
 import { url } from "../../components/App/App"
