+++ client/src/pages/DogUpdateForm/DogUpdateForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../DogProfileForm/DogProfileForm.css'
+import '../DogCreeationForm/DogCreationForm.css'
 import axios from 'axios'
 import env from 'react-dotenv'
 import { url } from "../../components/App/App"
